+++ src/app/orders/shared/order.service.ts
@@ -1,15 +1,16 @@
 import {Injectable} from '@angular/core';
 import {AngularFirestore} from '@angular/fire/firestore';
-import {Observable} from 'rxjs';
+import {from, Observable} from 'rxjs';
 import {Order} from "./order.model";
-import {map} from "rxjs/operators";
+import {map, switchMap, tap} from "rxjs/operators";
 
 @Injectable({
   providedIn: 'root'
 })
 export class OrderService {
 
-  constructor(private db: AngularFirestore) { }
+  constructor(private db: AngularFirestore) {
+  }
 
   getOrders(): Observable<Order[]> {
     return this.db.collection<Order>('Orders')
@@ -24,13 +25,29 @@ export class OrderService {
                   id: action.payload.doc.id,
                   name: data.name
                 };
-            })
+              })
           })
       );
   }
 
-  deleteOrder(id: string): Promise<void> {
-    return this.db.doc<Order>('Orders/' + id)
-      .delete();
+  deleteOrder(id: string): Observable<void> {
+    return this.db.doc<Order>('Orders/recfsearfc' + id)
+      .get()
+      .pipe(
+        tap(orderDocument => {
+          //debugger;
+        }),
+        switchMap(orderDocument => {
+          if (!orderDocument || !orderDocument.data()) {
+            window.alert('Order does not exist or contains no data.');
+            throw new Error('Order not found');
+          } else {
+            return from(
+              this.db.doc<Order>('Orders/' + id)
+                .delete()
+            )
+          }
+        })
+      )
   }
 }
